Separate SourceLookupProvider
